Fixed "Quest Stage" Perk effects
They are two-byte values, not one-byte values, and they're followed by two padding bytes.

+++ native/src/DovahKit/dovah/forms/structs/perk_effect.h
@@ -27,7 +27,7 @@ namespace dovah::loaded_forms::structs {
 
             struct quest {
                form_reference_t quest;
-               uint8_t          stage = 0;
+               uint16_t         stage = 0;
             };
             struct ability {
                form_reference_t spell;
